data_server/requirements.txt
fastapi
uvicorn[standard]
yt-dlp
redis
python-dotenv
aiohttp